Cart Management minor change

+++ src/main/java/com/sorted/portal/assisting/beans/CartItems.java
@@ -2,12 +2,14 @@ package com.sorted.portal.assisting.beans;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CartItems {
 
 	private String product_name;
 	private String product_code;
-	private Long selling_price;
+	private BigDecimal selling_price;
 	private Long quantity;
 	private boolean secure_item;
 	private Integer current_status;
